refactor: 페이지 분리 및 폴더 구조 정리
