+++ src/components/Main/ProductDetailPage/index.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function ProductDetailPage() {
+  return <div>ProductDetailPage</div>;
+}
+
+export default ProductDetailPage;
