+++ src/components/Main/ShoppingCartPage/index.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function ShoppingCartPage() {
+  return <div>ShoppingCartPage</div>;
+}
+
+export default ShoppingCartPage;
