+++ src/components/Main/index.jsx
@@ -1,13 +1,9 @@
 import React from "react";
 import styled from "styled-components";
-import GridList from "../GridList";
-import ProductCard from "../ProductCard";
 
-const dummy = {
-  thumbnail: "https://cdn-mart.baemin.com/goods/85/1537405626217m0.jpg",
-  name: "PET보틀-정사각(420ml)",
-  price: 43400,
-};
+import ProductListPage from "./ProductListPage";
+import ProductDetailPage from "./ProductDetailPage";
+import ShoppingCartPage from "./ShoppingCartPage";
 
 const Container = styled.main`
   max-width: 1320px;
@@ -19,22 +15,9 @@ const Container = styled.main`
 function Main() {
   return (
     <Container>
-      <GridList>
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-        <ProductCard productInfo={dummy} />
-      </GridList>
+      <ProductListPage />
+      <ProductDetailPage />
+      <ShoppingCartPage />
     </Container>
   );
 }
